Refactor Hero component - Move static data to heroData module - Refactor component
