+++ src/data/heroData.js
@@ -1,0 +1,23 @@
+import { faFacebookSquare } from '@fortawesome/free-brands-svg-icons/faFacebookSquare';
+import { faGithubSquare } from '@fortawesome/free-brands-svg-icons/faGithubSquare';
+import { faLinkedin } from '@fortawesome/free-brands-svg-icons/faLinkedin';
+
+const heroIcons = [
+  {
+    ariaLabel: 'Visit Facebook page',
+    icon: faFacebookSquare,
+    url: 'https://www.facebook.com/gunnarjohnson',
+  },
+  {
+    ariaLabel: 'Visit GitHub page',
+    icon: faGithubSquare,
+    url: 'https://github.com/gunnarjohnson',
+  },
+  {
+    ariaLabel: 'Visit LinkedIn page',
+    icon: faLinkedin,
+    url: 'https://www.linkedin.com/in/gunnarmjohnson',
+  },
+];
+
+export default { heroIcons };
